nails-marketplace/backend/requirements.txt
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
billiard==4.2.2
celery==5.4.0
certifi==2025.10.5
cffi==2.0.0
channels==4.2.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.1.2
Django==5.2.8
django-allauth==65.13.0
django-appconf==1.2.0
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-environ==0.12.0
django-extensions==3.2.3
django-filter==24.3
django-imagekit==5.0.0
djangorestframework==3.15.2
dnspython==2.8.0
drf-yasg==1.21.7
geographiclib==2.1
geopy==2.4.1
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
inflection==0.5.1
iniconfig==2.3.0
kombu==5.6.0
mercadopago==2.2.3
msgpack==1.1.2
oauthlib==3.3.1
packaging==25.0
pilkit==3.0
pillow==12.0.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pymongo==4.10.1
pyOpenSSL==25.3.0
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
redis==5.2.1
reportlab==4.2.5
requests==2.32.5
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
zope.interface==8.1
